requirements.txt
ruamel.yaml
aiodocker
docker
loko-client
boto3
kubernetes
click
paramiko
loguru
